Performance Improvements 1.11.2
- Fixed potential crash in conveyor Reachable cache logic

It seemed to be a race condition which I could not reproduce, but found in the code. Since I could not verify the fix an error handler is added to silently log any such exception and just call the original method instead. This way the server won't crash and the result will also be correct.