Multigrid Projector 0.6.2
- Torch configuration option: Set preview block visuals on server side (required by mods depending on preview block transparency) [DEFAULT: OFF]
- Enabled "set preview block visuals" permanently on Dedicated Server to avoid the same issue by default (configuration may be added later)
- Minor optimization to setting the preview block visuals on client side (not setting them redundantly if the whole subgrid is unchanged)

Set the new Torch configuration option if you use the ToolCore mod. It is required for ToolCore welders to work until a proper fix is implemented in the mod. Enabling this option has a minor server performance impact during welding, therefore it should be disabled once the mod is fixed.