Toolbar Manager 1.5.1
- Using character profiles for build cockpit

This is to prevent players from thinking they lost all of their character profiles while trying to use the cockpit build mode (Ctrl-G).

This feature is enabled by default, but will not be in effect if there is any profiles aready defined for the build cockpit.

To merge your build cockpit profiles into your character profiles open the `%AppData%\SpaceEngineers\ToolbarManager` directory in `Explorer` and move all files from the `BuildCockpit` to the `Character` folder.